backend/requirements/base.txt
# Core
Django==5.1.*
djangorestframework==3.15.*
django-filter==24.3

# Auth
djangorestframework-simplejwt==5.3.*

# Database
mysqlclient==2.2.*

# Async / Tasks
celery[redis]==5.4.*
django-celery-beat==2.7.*
redis==5.2.*

# WebSocket
channels==4.1.*
channels_redis==4.2.*

# Payments
stripe==11.*

# AI
openai==1.*

# Security / Encryption
cryptography==43.*
python-decouple==3.8

# Utilities
Pillow==11.*
phonenumbers==8.13.*

# CORS
django-cors-headers==4.6.*

# HTTP client
httpx==0.28.*
